Time: 3 ms (28.00%) | Memory: 95.8 MB (30.37%) - LeetSync

+++ 4-median-of-two-sorted-arrays/median-of-two-sorted-arrays.cpp
@@ -8,7 +8,7 @@ public:
         if(size % 2 == 1){
             ans = nums1[size/2];
         }else{
-            ans = (nums1[size/2 - 1] + nums1[size/2])/ 2.0;
+            ans = (double)(nums1[size/2 - 1] + nums1[size/2])/ 2;
         }
         return ans;
     }
